Store name in Cyrillic script in :name_ru
The names on EP are currently displayed in latin. We will store
the latin script transliterations in :name. As a step towards this
we are moving the cyrillic spellings to :name_ru.

(I have chose _ru over _be as the url indicates that these are
Russian spellings: http://house.gov.by/ru/deputies-ru/)