Add ContactInfo class to handle member contact info
This class does all the work of extracting email addresses and
web addresses from the contact fields in the member pages.

It provides two methods: #email_addresses and #web_addresses. Each
returns an array of relevant addresses found in the contact info
box (passed to the initializer as a noko node set).